Make new popups folder and update import paths
