+++ src/components/addNewRequest/NewRequest.tsx
@@ -3,9 +3,9 @@ import { useNavigate } from 'react-router-dom';
 import Header from '../Header';
 import ReceiptInfoForm from '../receiptInfoForm/ReceiptInfoForm';
 import ReceiptPreview from '../receiptPreview/ReceiptPreview';
-import FormScanningPopup from '../addRequestPopups/FormScanningPopup';
-import FormSubmittedPopup from '../addRequestPopups/FormSubmittedPopup';
-import FormNotAcceptedPopup from '../addRequestPopups/FormNotAcceptedPopup';
+import FormScanningPopup from '../popups/addRequestPopups/FormScanningPopup';
+import FormSubmittedPopup from '../popups/addRequestPopups/FormSubmittedPopup';
+import FormNotAcceptedPopup from '../popups/addRequestPopups/FormNotAcceptedPopup';
 import { useError } from '../../context/errorContext';
 import { useAuth } from '../../context/authContext';
 import { postNewRequest } from '../../services/dataServices/claimsRequest';
