+++ src/components/popups/addRequestPopups/FormNotAcceptedPopup.tsx
@@ -1,40 +1,41 @@
 import React from 'react';
 import './Popups.css';
-import CheckCircleOutlinedIcon from '@mui/icons-material/CheckCircleOutlined';
-import ActionButton from '../actionButton/ActionButton';
+import ErrorOutlineOutlinedIcon from '@mui/icons-material/ErrorOutlineOutlined';
+import ActionButton from '../../actionButton/ActionButton';
 
-interface FormSubmittedPopupProps {
-    onViewHistory: () => void;
-    onClose: () => void;
+interface FormNotAcceptedPopupProps {
+    onReview: () => void;
+    onResubmit: () => void;
 }
 
-const FormSubmittedPopup: React.FC<FormSubmittedPopupProps> = ({ onViewHistory, onClose }) => {
+const FormNotAcceptedPopup: React.FC<FormNotAcceptedPopupProps> = ({ onReview, onResubmit }) => {
     return (
         <div className="popup-overlay">
-            <div className="popup-content-add-request submitted">
+            <div className="popup-content-add-request not-accepted">
                 <div className="text">
                     <div className="popup-icon">
-                        <CheckCircleOutlinedIcon style={{ color: '#86B62C' }} />
+                        <ErrorOutlineOutlinedIcon style={{ color: '#F6772C' }} />
                     </div>
                     <div className="popup-main-text" style={{ color: '#4C4E54' }}>
-                        Form Submitted
+                        Form not accepted
                     </div>
                     <div className="popup-secondary-text" style={{ color: '#626469' }}>
-                        The request has been accepted. View request in Claim History.
+                        There was an error in submitting your form.
                     </div>
                 </div>
                 <div className="popup-actions">
                     <ActionButton
                         className="popup-button primary-button"
-                        handleEvent={onViewHistory}
+                        handleEvent={onReview}
                         type="button"
-                        placeholder="View Claim History"
+                        placeholder="Review Request"
                     />
+
                     <ActionButton
                         className="popup-button secondary-popup-button"
-                        handleEvent={onClose}
+                        handleEvent={onResubmit}
                         type="button"
-                        placeholder="Close"
+                        placeholder="Resubmit"
                     />
                 </div>
             </div>
@@ -42,4 +43,4 @@ const FormSubmittedPopup: React.FC<FormSubmittedPopupProps> = ({ onViewHistory, 
     );
 };
 
-export default FormSubmittedPopup;
+export default FormNotAcceptedPopup;
