+++ src/components/popups/changePasswordPopup/ChangePasswordPopup.css
@@ -53,19 +53,6 @@
     margin-bottom: 10px !important;
 } 
 
-.submit-button-change-password {
-    background-color: #1CA8DD !important;
-    color: white !important;
-    font-family: 'LLCirc' !important;
-    font-weight: 100 !important;
-    box-shadow: none !important;
-    border-radius: 4px !important;
-    font-size: 14px !important;
-    margin-right: 17px !important;
-    text-transform: none !important;
-    width: 90px;
-} 
-
 .MuiDialogContent-root {
     border: none !important;
 }
