+++ src/pages/claim_details/ClaimDetails.tsx
@@ -5,9 +5,9 @@ import { useError } from '../../context/errorContext';
 import Header from '../../components/Header';
 import ReceiptView from '../../components/receiptView/ReceiptView';
 import ReceiptPreview from '../../components/receiptPreview/ReceiptPreview';
-import AcceptPopup from '../../components/viewFormPopups/AcceptPopup';
-import DeclinePopup from '../../components/viewFormPopups/DeclinePopup';
-import EditPopup from '../../components/viewFormPopups/EditPopup';
+import AcceptPopup from '../../components/popups/viewFormPopups/AcceptPopup';
+import DeclinePopup from '../../components/popups/viewFormPopups/DeclinePopup';
+import EditPopup from '../../components/popups/viewFormPopups/EditPopup';
 import UserTitle from '../../components/userTitle/UserTitle';
 import LoadingScreen from '../../components/loadingScreen/LoadingScreen';
 import { downloadClaimImages, updateClaimStatus } from '../../services/dataServices/claimsHistory';
