+++ src/pages/user_profile/UserProfile.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import ProfileTable from '../../components/profile/ProfileTable';
 import UserTitle from '../../components/userTitle/UserTitle';
 import Header from '../../components/Header';
-import ChangePasswordPopup from '../../components/profile/ChangePasswordPopup';
+import ChangePasswordPopup from '../../components/popups/changePasswordPopup/ChangePasswordPopup';
 import { fetchProfile } from '../../utils/userProfileUtils';
 import { useAuth } from '../../context/authContext';
 import { useError } from '../../context/errorContext';
